fix webseeds > 2GB (ish)
git-svn-id: http://svn.vuze.com/public/client@25014 c2c1b7af-1477-405b-86ee-8aefd1fc4dd2

+++ trunk/azureus2/src/com/aelitis/azureus/plugins/extseed/impl/getright/ExternalSeedReaderGetRight.java
@@ -289,7 +289,7 @@ ExternalSeedReaderGetRight
 	{	
 		setReconnectDelay( RECONNECT_DEFAULT, false );
 		
-		long	request_start 	= start_piece_number * piece_size + start_piece_offset;
+		long	request_start 	= start_piece_number * (long)piece_size + start_piece_offset;
 		int		request_length	= length;
 		
 		if ( http_downloaders.length == 1 ){
